refactor(indexer): move computeRelevance to utils and adjust ENS thresholds
- Move computeRelevance function from constants.ts to utils.ts
- Update imports in event handlers (delegation, transfer, voting)
- Adjust ENS delegation thresholds: medium 10k, high 100k
- Adjust ENS vote thresholds: low 1k, medium 10k, high 100k
