+++ apps/indexer/src/lib/utils.ts
@@ -1,5 +1,8 @@
 import * as chains from "viem/chains";
 
+import { DaoIdEnum, FeedEventRelevanceEnum, FeedEventTypeEnum } from "./enums";
+import { RELEVANCE_THRESHOLDS } from "./constants";
+
 /**
  * Calculates the absolute difference between two numbers
  */
@@ -29,4 +32,38 @@ export function max(...values: bigint[]): bigint {
 
 export function getChain(chainId: number): chains.Chain | undefined {
   return Object.values(chains).find((chain) => chain.id === chainId);
+}
+
+/**
+ * Computes the relevance level for a feed event based on the value and DAO thresholds.
+ * If thresholds are not configured for the DAO, returns "none".
+ */
+export function computeRelevance(
+  daoId: DaoIdEnum,
+  eventType:
+    | FeedEventTypeEnum.TRANSFER
+    | FeedEventTypeEnum.DELEGATION
+    | FeedEventTypeEnum.VOTE,
+  value: bigint,
+): FeedEventRelevanceEnum {
+  const thresholds = RELEVANCE_THRESHOLDS[daoId];
+  if (!thresholds) {
+    return FeedEventRelevanceEnum.NONE;
+  }
+
+  const eventThresholds = thresholds[eventType];
+  if (!eventThresholds) {
+    return FeedEventRelevanceEnum.NONE;
+  }
+
+  if (value >= eventThresholds.high) {
+    return FeedEventRelevanceEnum.HIGH;
+  }
+  if (value >= eventThresholds.medium) {
+    return FeedEventRelevanceEnum.MEDIUM;
+  }
+  if (value >= eventThresholds.low) {
+    return FeedEventRelevanceEnum.LOW;
+  }
+  return FeedEventRelevanceEnum.NONE;
 }
